requirements.txt
Jinja2>=3.0.1,==3.*
Markdown>=3.3.4,==3.*
pydantic>=1.8.2,==1.*
PyYAML>=5.4.1,==5.*
beautifulsoup4>=4.9.3,==4.*
python-frontmatter>=1.0.0,==1.*
